Monika Real Poem Challenge Submod For MAS
Extract this zip file into the ''Submods'' folder that should exist inside the ''game'' folder conatined in the ''DDLC MAS'' folder.

This Script adds the Real Poem Challenge event to MAS, this event is from my own mod DDLC REAL FRIENDS, this script also adds extra dialogues, intros, and ''I feell'' choices, you need to wait for her to bring up the Real Poem Challenge and then make her repeat the topic. If you don't like this submod please delete it from MAS.

There might be some small typos here and there, this was purely for fun and to share with some friends so please don't bash me for it.